refactor: move layout pages to layout folder
